+++ wooqi/__init__.py
@@ -7,4 +7,4 @@
 """
 Wooqi version
 """
-__version__ = "4.0.2"
+__version__ = "5.0.0"