+++ wooqi/src/__init__.py
@@ -10,5 +10,4 @@ Define plugin global variable
 import logging
 
 global_var = {}
-logger_gv = logging.getLogger()
-logger_gv.init = False
+logger = logging.getLogger(__package__)